# BACKEND/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API Laravel CRUD",
        "description": "Documentação da API Laravel CRUD",
        "contact": {
            "email": "seu-email@dominio.com"
        },
        "version": "1.0.0"
    },
    "paths": {
        "/api/clients": {
            "get": {
                "tags": [
                    "Clients"
                ],
                "summary": "Listar todos os clientes",
                "operationId": "d00d20521584ab5da83e088abb7aded2",
                "responses": {
                    "200": {
                        "description": "Lista de clientes retornada com sucesso"
                    }
                }
            },
            "post": {
                "tags": [
                    "Clients"
                ],
                "summary": "Criar um novo cliente",
                "operationId": "f1e86e5b79fedfa6593067977292b13b",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "name",
                                    "cpf",
                                    "gender",
                                    "email"
                                ],
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "example": "Maria Silva"
                                    },
                                    "cpf": {
                                        "type": "string",
                                        "example": "123.456.789-00"
                                    },
                                    "gender": {
                                        "type": "string",
                                        "example": "feminino"
                                    },
                                    "email": {
                                        "type": "string",
                                        "example": "maria@email.com"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Cliente criado com sucesso"
                    },
                    "422": {
                        "description": "Erro de validação"
                    }
                }
            }
        },
        "/api/clients/{client}": {
            "get": {
                "tags": [
                    "Clients"
                ],
                "summary": "Exibir um cliente específico",
                "operationId": "e1b94279234de95a7144b78642cd2e45",
                "parameters": [
                    {
                        "name": "client",
                        "in": "path",
                        "description": "ID do cliente",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Cliente encontrado com sucesso"
                    },
                    "404": {
                        "description": "Cliente não encontrado"
                    }
                }
            },
            "put": {
                "tags": [
                    "Clients"
                ],
                "summary": "Atualizar um cliente",
                "operationId": "0c9a2108c7f6120cd8c12ece56a85b74",
                "parameters": [
                    {
                        "name": "client",
                        "in": "path",
                        "description": "ID do cliente",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "name",
                                    "cpf",
                                    "gender",
                                    "email"
                                ],
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "example": "Maria Silva"
                                    },
                                    "cpf": {
                                        "type": "string",
                                        "example": "123.456.789-00"
                                    },
                                    "gender": {
                                        "type": "string",
                                        "example": "feminino"
                                    },
                                    "email": {
                                        "type": "string",
                                        "example": "maria@email.com"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Cliente atualizado com sucesso"
                    },
                    "422": {
                        "description": "Erro de validação"
                    }
                }
            },
            "delete": {
                "tags": [
                    "Clients"
                ],
                "summary": "Excluir um cliente",
                "operationId": "2d450bfd555b490e7e9cdd602d1db8e3",
                "parameters": [
                    {
                        "name": "client",
                        "in": "path",
                        "description": "ID do cliente",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Cliente excluído com sucesso"
                    },
                    "404": {
                        "description": "Cliente não encontrado"
                    }
                }
            }
        },
        "/api/products": {
            "get": {
                "tags": [
                    "Products"
                ],
                "summary": "Listar todos os produtos",
                "operationId": "1bfaa78d1c2c3848ab8165c5dadcad3e",
                "responses": {
                    "200": {
                        "description": "Lista de produtos retornada com sucesso"
                    }
                }
            },
            "post": {
                "tags": [
                    "Products"
                ],
                "summary": "Criar um novo produto",
                "operationId": "e8185e307706e2cd84bbcf5171e2f09d",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "name",
                                    "color",
                                    "price"
                                ],
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "example": "Notebook Gamer"
                                    },
                                    "color": {
                                        "type": "string",
                                        "example": "Preto"
                                    },
                                    "price": {
                                        "type": "string",
                                        "example": "3499.99"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Produto criado com sucesso"
                    },
                    "422": {
                        "description": "Erro de validação"
                    }
                }
            }
        },
        "/api/products/{product}": {
            "get": {
                "tags": [
                    "Products"
                ],
                "summary": "Exibir um produto específico",
                "operationId": "05f6ffcb49f47b4f4fc1a1c21ed60011",
                "parameters": [
                    {
                        "name": "product",
                        "in": "path",
                        "description": "ID do produto",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Produto retornado com sucesso"
                    },
                    "404": {
                        "description": "Produto não encontrado"
                    }
                }
            },
            "put": {
                "tags": [
                    "Products"
                ],
                "summary": "Atualizar um produto",
                "operationId": "c474d1318e5b8c3de1f4d52b6053111f",
                "parameters": [
                    {
                        "name": "product",
                        "in": "path",
                        "description": "ID do produto",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "example": "Notebook Atualizado"
                                    },
                                    "color": {
                                        "type": "string",
                                        "example": "Cinza"
                                    },
                                    "price": {
                                        "type": "string",
                                        "example": "2999.99"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Produto atualizado com sucesso"
                    },
                    "422": {
                        "description": "Erro de validação"
                    }
                }
            },
            "delete": {
                "tags": [
                    "Products"
                ],
                "summary": "Deletar um produto",
                "operationId": "1a0a353d0842ff4f9347bb66ec53e3ea",
                "parameters": [
                    {
                        "name": "product",
                        "in": "path",
                        "description": "ID do produto",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Produto deletado com sucesso"
                    },
                    "404": {
                        "description": "Produto não encontrado"
                    }
                }
            }
        },
...
    },
...
}